refactor(studio): lift the carve out of the FX section file
`propertyPanelFxSection.tsx` was 992 lines against the studio's 600-line
cap. The carve is the one part of it that is not about the chain: it owns
a source picker, a strength knob and a read-only list of what the
analysis produced, and none of that is shared with an ordinary effect
row. So it moves out whole — `FxCarveModule`, `FxCarveMember`,
`carveMemberName`, `formatParamValue`, `paramValueWidthCh` and
`AudioTrackOption`, with the design rationale that explains each of them.

Pure move: no behaviour change, no rendered-audio change. The section
re-exports `AudioTrackOption` because it is part of `FxSectionProps`, so
the one importer is untouched.

Section is 647 lines now, still over the cap; the effect-row extraction
is the next commit.

Studio suite unchanged at 3674 passing, 18 todo, 1 file skipped.
